PDF file with Poster and description
Project were created on the "Flack" - flash-based map interface for
allocating ProtoGENI and other supported GENI resources. There were
reserved one slice with four nodes. Two of them are two LINCs installed
on Ubuntu 12.04. The third one is a Network Manager. It is hosting the
GUI to display all LINC’s OpenFlow parameters. Also, Network Manger runs
NETCONF over SSH Client that is needed to establish sessions with
NETCONF Server on LINCs and retrieve XML configuration schemes. NETCONF
over SSH Client/Server is a YANG-Based Unified Modular Automation Tool,
that is basically plug-in to the OpenFlow framework, and it is used to
push/pull configurations to /from switches and clients.